apps/api/requirements.txt
# ─── Core ─────────────────────────────────────────────────────
Django==4.2.16
django-ninja==1.3.0
django-cors-headers==4.4.0
django-ratelimit==4.1.0
django-environ==0.11.2

# ─── Auth / JWT ───────────────────────────────────────────────
djangorestframework-simplejwt==5.3.1
PyJWT==2.9.0
email-validator==2.2.0

# ─── Database / Cache ────────────────────────────────────────
psycopg[binary]==3.2.3
redis==5.0.8
hiredis==3.0.0

# ─── Servidor WSGI de produção ────────────────────────────────
gunicorn==22.0.0

# ─── Workflows ────────────────────────────────────────────────
temporalio==1.7.1

# ─── Observabilidade / Logging ────────────────────────────────
structlog==24.4.0
sentry-sdk[django]==2.14.0

# ─── HTTP client (para chamadas à Groq API) ───────────────────
httpx==0.27.2

# ─── Dev / Testing ────────────────────────────────────────────
pytest==8.3.3
pytest-django==4.9.0
pytest-cov==5.0.0
ruff==0.6.9
